pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uncoded"
version = "0.0.0"
description = "Here be dragons."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Alistair Miles", email = "alimanfoo@googlemail.com" },
]
dependencies = [
    "pyyaml>=6.0",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

[project.scripts]
uncoded = "uncoded.cli:main"

[project.urls]
Homepage = "https://github.com/alimanfoo/uncoded"
Repository = "https://github.com/alimanfoo/uncoded"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
